requirements.txt
# local package
-e .

# external requirements
click
coverage
awscli
flake8
python-dotenv>=0.5.1
geopandas
requests
ipykernel
cenpy
